backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
supabase==2.5.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
